# api/swagger/swagger.v1.yaml
openapi: "3.0.2"
info:
  title: Resources Object CRUD Operations
  description: Resources CRUD Operations
  version: "1.0"
  contact:
    name: Karthikeyan
    email: karthikeyankc91@gmail.com
    url: https://github.com/karthikeyankc91

servers:
  - url: http://localhost:3000/

paths:
  /v1/resources:
    get:
      tags:
        - CRUD Operations
      responses:
        200:
          description: OK
    post:
      summary: Create a resource
      description: |
        # Create Resource
        * Authorization Basic
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceRequest"
            example:
              metadata:
                name: boon123
              spec:
                name: Karthikeyan KC
                age: 30
                nationality: Indian
      tags:
        - CRUD Operations
      responses:
        201:
          description: Successful

  /v1/resources/{name}:
    get:
      summary: Gets a resource by name
      description: |
        # Get a resource by name
        * Authentication: Basic
      parameters:
        - $ref: "#/components/parameters/NamePath"
      tags:
        - CRUD Operations
      responses:
        200:
          description: OK
    patch:
      summary: Patch a resource
      description: |
        # Patch resource
        * Authentication Basic
      parameters:
        - $ref: "#/components/parameters/NamePath"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourcePatchRequest"
            example:
              spec:
                age: 31
                nationality: Indian
      tags:
        - CRUD Operations
      responses:
        200:
          description: OK
    delete:
      summary: Deletes a resource by name
      description: |
        # Get a resource by name
        * Authentication: Basic
      parameters:
        - $ref: "#/components/parameters/NamePath"
      tags:
        - CRUD Operations
      responses:
        200:
          description: OK

components:
  parameters:
    NamePath:
      name: name
      description: What is the name?
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Name"

  schemas:
    Name:
      type: string
      description: Name of the resource

    Resource:
      title: Resource Definition
      description: Definition of a resource Object
      type: object
      properties:
        metadata:
          $ref: "#/components/schemas/Metadata"
        spec:
          $ref: "#/components/schemas/Spec"

    Metadata:
      title: Metadata of the resource
      description: The name of the resource object
      type: object
      properties:
        name:
          type: string
          description: The name of the resource object

    Spec:
      title: Spec of the resource
      description: The specifications of the resource object
      type: object
      properties:
        name:
          type: string
          description: The name of the resource object
        age:
          type: number
          minimum: 1
          maximum: 150
          description: The name of the resource object
        nationality:
          type: string
          description: The name of the resource object
      required:
        - name
        - age
        - nationality

    PatchSpec:
      title: Spec of the resource
      description: The specifications of the resource object
      type: object
      properties:
        age:
          type: number
          minimum: 1
          maximum: 150
          description: The name of the resource object
        nationality:
          type: string
          description: The name of the resource object
      required:
        - age
        - nationality

    ResourceRequest:
      title: Resource request for creation
      description: Resource request for creation
      type: object
      properties:
        metadata:
          $ref: "#/components/schemas/Metadata"
        spec:
          $ref: "#/components/schemas/Spec"

    ResourcePatchRequest:
      title: Resource request for creation
      description: Resource request for creation
      type: object
      properties:
        spec:
          $ref: "#/components/schemas/PatchSpec"